LUIS-Bot/requirements.txt
aiohttp
botbuilder-core>=4.7.0
botbuilder-dialogs>=4.7.0
botbuilder-ai>=4.7.0
botbuilder-azure>=4.7.0
